feat(protocol): improve buffer management in Encode functions and add buffer validation in putEncodeBuffer
feat(storage): remove test helper functions for setting and expiring value objects
